Reduce code duplication by introducing HybridRecordReceiver
